transformers/image-captioning/requirements.txt
torch
torchvision
transformers
gradio
Pillow